Change WorldMapTooltip drawlayer to default value
Changing the FrameStrata of WorldMapFrame also changes the strata
of its child elements. This turns the WorldMapTooltip from "TOOLTIP"
down to "HIGH" which places it below some elements of other addons.

The game default of the WorldMapTooltip is "TOOLTIP" and I don't see
a reason why it should be set to only "HIGH" as it just breaks things.

By this commit, the FrameStrata of the WorldMapTooltip is set back to
the game default value: "TOOLTIP".

Fixes: https://github.com/shagu/pfQuest/issues/4